requirements.txt
asgiref==3.9.1
django==4.2.24
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1